+++ src/objects/Object.svelte
@@ -20,6 +20,7 @@ Core component for direct-manipulation objects. A drag&drop source.
     }
 
     export let id: string = "";
+    export let tip: string = "";
     export let label: string;
     export let selected: boolean = false;
     export let conflicted: boolean;
@@ -85,6 +86,7 @@ Core component for direct-manipulation objects. A drag&drop source.
     tabindex="-1"
     draggable="true"
     role="option"
+    title={tip}
     aria-label={label}
     aria-selected={selected}
     on:click={onClick}